Updated to use QFJ with BigDecimals that addresses fix for QFJ-300 (http://www.quickfixj.org/jira/browse/QFJ-300)
git-svn-id: file:///space/source.marketcetera.org@1718 501a4a96-914b-0410-8454-f5988a4ca596

+++ root/trunk/source/core/src/test/java/org/marketcetera/quickfix/FIXVersionsTest.java
@@ -45,7 +45,7 @@ public class FIXVersionsTest extends TestCase {
 
     }
 
-    private Message createNOSHelper(FIXVersion version, String inSymbol, double qty, char inSide)
+    private Message createNOSHelper(FIXVersion version, String inSymbol, int qty, char inSide)
     {
         long suffix = System.currentTimeMillis();
         Message newSingle = version.getMessageFactory().newMarketOrder("123"+suffix, inSide, new BigDecimal(qty), new MSymbol(inSymbol),
